refactor: move routing stuff into subfolder routing
